sync: shini - long to int (prerequisit for 64-bit)

+++ pol-core/bscript/berror.cpp
@@ -13,8 +13,8 @@ Notes
 #include "impstr.h"
 #include "contiter.h"
 
-unsigned long BError::creations_ = 0;
-unsigned long BError::creations()
+unsigned int BError::creations_ = 0;
+unsigned int BError::creations()
 {
     return creations_;
 }
