+++ pol-core/plib/mapwriter.h
@@ -32,23 +32,23 @@ public:
     void OpenExistingFiles(const string& realm_name);
     void WriteConfigFile();
 
-    unsigned long NextSolidIndex();
-    unsigned long NextSolidOffset();
+    unsigned int NextSolidIndex();
+    unsigned int NextSolidOffset();
 
-    unsigned long NextSolidx2Offset();
+    unsigned int NextSolidx2Offset();
 
     void AppendSolid( const SOLIDS_ELEM& solid );
     void AppendSolidx2Elem( const SOLIDX2_ELEM& elem );
-    void SetSolidx2Offset( unsigned short x_base, unsigned short y_base, unsigned long offset );
+    void SetSolidx2Offset( unsigned short x_base, unsigned short y_base, unsigned int offset );
 
     unsigned width() const { return _width; }
     unsigned height() const { return _height; }
 private:
     std::fstream::pos_type total_size();
     
-    unsigned long total_solid_blocks();
-    unsigned long total_blocks();
-    unsigned long total_maptile_blocks();
+    unsigned int total_solid_blocks();
+    unsigned int total_blocks();
+    unsigned int total_maptile_blocks();
 
     void FlushBaseFile();
     void FlushMapTileFile();
@@ -58,7 +58,7 @@ private:
     unsigned short _width;
     unsigned short _height;
     fstream _ofs_base;
-    long _cur_mapblock_index;
+    int _cur_mapblock_index;
     MAPBLOCK _block;
 
     fstream _ofs_solidx1;
@@ -66,7 +66,7 @@ private:
     fstream _ofs_solids;
 
     fstream _ofs_maptile;
-    long _cur_maptile_index;
+    int _cur_maptile_index;
     MAPTILE_BLOCK _maptile_block;
 };
