+++ pol-core/pol/decay.h
@@ -11,5 +11,5 @@ Notes
 void decay_items();
 void decay_single_zone();
 
-extern long cycles_per_decay_worldzone;
-extern long cycles_until_decay_worldzone;
+extern int cycles_per_decay_worldzone;
+extern int cycles_until_decay_worldzone;
